rust: clippy-clean workspace
Auto-fix pass plus manual cleanup: bind! codegen emits From impls
without the Copy clone, api Into impls flipped to From, slice params
replace &Vec, dead verbatim-port items carry explicit allows. Style
lints that fight the vanilla ports (index loops, negated float
comparisons keeping vanilla NaN semantics, JNI argument counts) are
allowed at crate level with a justification comment. Workspace
resolver pinned to 2; macros crate declares proc-macro = true.
cargo clippy 0 warnings, cargo test 153/153.

+++ rust/mod/src/nav_cache_jni.rs
@@ -42,7 +42,7 @@ pub extern "system" fn Java_me_apika_apikaprobe_RustBridge_navFillSection<'local
 ) {
     let Some(ptr) = env.get_direct_buffer_address(&cell_data).ok() else { return };
     let Some(cap) = env.get_direct_buffer_capacity(&cell_data).ok() else { return };
-    let raw = unsafe { slice::from_raw_parts(ptr, cap as usize) };
+    let raw = unsafe { slice::from_raw_parts(ptr, cap) };
     let id = SectionId { chunk_x, section_y, chunk_z };
     nav_cache_storage::fill_section(id, raw);
 }
